chore: split tasks test again
